# compiler: reuse derived item properties from last compilation 