--- a/compiler/src/Config.hs
+++ b/compiler/src/Config.hs
@@ -84,7 +84,7 @@ readConfig = decodeYamlFile
 data ViewerConfig = ViewerConfig
   { galleryTitle :: String
   , tagCategories :: [String]
-  } deriving (Generic, ToJSON, Show)
+  } deriving (Generic, ToJSON, FromJSON, Show)
 
 viewerConfig :: GalleryConfig -> ViewerConfig
 viewerConfig GalleryConfig{galleryTitle, tagCategories} = ViewerConfig galleryTitle tagCategories