… compiler/src/Resource.hs