A benchmark on an already bulit gallery with ~600 pictures shows a ~90% speedup:

Before:

    Time (mean ± σ):      2.879 s ±  0.125 s    [User: 14.686 s, System: 5.511 s]
    Range (min … max):    2.774 s …  3.203 s    10 runs

After:

    Time (mean ± σ):     289.5 ms ±  15.1 ms    [User: 596.1 ms, System: 359.3 ms]
    Range (min … max):   272.8 ms … 323.0 ms    10 runs

---

GitHub: closes #97
\+ some refactoring

---

@OzoneGrif please test and compare perfs before/after on your big gallery over that slow SMB share